Extract SC integration tests to a separate file
